requirements.txt
django
djangorestframework==3.11.0
django-rest-auth==0.9.5
psycopg2
requests==2.24.0